backend/services/gradimento-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
asyncpg==0.30.0
databases==0.8.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

